Moved init up a level so as to share common functions between queries and mutations, will have to do the same across branches. Also, Updated person queries and tests as per discussion with alastair, have to do the same for document
